chore: 변수명 및 폴더링 정리
