+++ DevLog/Resource/Localizable.xcstrings
@@ -447,7 +447,7 @@
       "localizations" : {
         "en" : {
           "stringUnit" : {
-            "state" : "needs_review",
+            "state" : "translated",
             "value" : "Saved web pages appear here."
           }
         },
